requirements.txt
fpdf2
pyttsx3
streamlit
pandas
matplotlib
opcua
